SudokuPermutatedCellsChromosome and SudokuOrderedCellsChromosome working with permutation initialisation
