+++ Sudoku.GeneticSharp/SudokuCellsChromosome.cs
@@ -34,16 +34,20 @@ public class SudokuCellsChromosome : SudokuChromosomeBase, ISudokuChromosome
 	{
 		var row = geneIndex / 9;
 		var col = geneIndex % 9;
+		/*
 		var cellIndex = (row, col);
 		//If a target mask exist and has a digit for the cell, we use it.
 		if (TargetSudokuGrid != null && TargetSudokuGrid.Cells[row][col] != 0)
 		{
 			return new Gene(TargetSudokuGrid.Cells[row][col]);
 		}
-		// otherwise we use a random digit amongts those permitted.
+		// otherwise we use a random digit amongst those permitted.
 		var rnd = RandomizationProvider.Current;
 		var targetIdx = rnd.GetInt(0, ExtendedMask[cellIndex].Count);
 		return new Gene(ExtendedMask[cellIndex][targetIdx]);
+		*/
+		Gene toReturn = new Gene(Permutations[row][col]);
+		return toReturn;
 	}
 
 	public override IChromosome CreateNew()
