+++ Sudoku.GeneticSharp/SudokuPermutationsChromosome.cs
@@ -9,12 +9,6 @@ namespace Sudoku.GeneticSharp;
 /// </summary>
 public class SudokuPermutationsChromosome : SudokuChromosomeBase, ISudokuChromosome
 {
-
-	/// <summary>
-	/// The list of row permutations accounting for the mask
-	/// </summary>
-	private IList<IList<IList<int>>> _targetRowsPermutations;
-
 	/// <summary>
 	/// This constructor assumes no mask
 	/// </summary>
@@ -36,7 +30,20 @@ public class SudokuPermutationsChromosome : SudokuChromosomeBase, ISudokuChromos
 	/// <param name="targetSudokuGrid">the target sudoku to solve</param>
 	/// <param name="extendedMask">The cell domains after initial constraint propagation</param>
 	/// <param name="length">The number of genes for the sudoku chromosome</param>
-	public SudokuPermutationsChromosome(SudokuGrid targetSudokuGrid, Dictionary<(int row, int col), List<int>> extendedMask, int length) : base(targetSudokuGrid, extendedMask, length) { }
+	public SudokuPermutationsChromosome(SudokuGrid targetSudokuGrid,
+		Dictionary<(int row, int col), List<int>> extendedMask, int length)
+		: base(targetSudokuGrid, extendedMask, length)
+	{
+	}
+	
+	public SudokuPermutationsChromosome(
+		SudokuGrid targetSudokuBoard,
+		Dictionary<(int x, int y), List<int>> extendedMask,
+		IList<(int row, int col)> objGeneToCellLookup,
+		List<int> objBaseLookupTable)
+		: base(targetSudokuBoard, extendedMask, 9)
+	{
+	}
 
 	/// <summary>
 	/// generates a chromosome gene from its index containing a random row permutation
@@ -46,11 +53,8 @@ public class SudokuPermutationsChromosome : SudokuChromosomeBase, ISudokuChromos
 	/// <returns>a gene generated for the index</returns>
 	public override Gene GenerateGene(int geneIndex)
 	{
-
-		var rnd = RandomizationProvider.Current;
-		//we randomize amongst the permutations that account for the target mask.
-		var permIdx = rnd.GetInt(0, TargetRowsPermutations[geneIndex].Count);
-		return new Gene(permIdx);
+		Gene toReturn = new Gene(PermutationsGenes[geneIndex]);
+		return toReturn;
 	}
 
 	public override IChromosome CreateNew()
@@ -98,7 +102,7 @@ public class SudokuPermutationsChromosome : SudokuChromosomeBase, ISudokuChromos
 
 		// we use a modulo operator in case the gene was swapped:
 		// It may contain a number higher than the number of available permutations. 
-		var perm = TargetRowsPermutations[rowIndex][permIDx % TargetRowsPermutations[rowIndex].Count].ToList();
+		var perm = GetRowsPermutations()[rowIndex][permIDx % GetRowsPermutations()[rowIndex].Count].ToList();
 		return perm;
 	}
 
@@ -112,146 +116,4 @@ public class SudokuPermutationsChromosome : SudokuChromosomeBase, ISudokuChromos
 	{
 		return (int)GetGene(rowIndex).Value;
 	}
-
-	/// <summary>
-	/// This method computes for each row the list of digit permutations that respect the target mask, that is the list of valid rows discarding columns and boxes
-	/// </summary>
-	/// <param name="SudokuGrid">the target sudoku to account for</param>
-	/// <returns>the list of permutations available</returns>
-	public IList<IList<IList<int>>> GetRowsPermutations()
-	{
-		if (TargetSudokuGrid == null)
-		{
-			return UnfilteredPermutations;
-		}
-
-		// we store permutations to compute them once only for each target Sudoku
-		if (!_rowsPermutations.TryGetValue(TargetSudokuGrid, out var toReturn))
-		{
-			// Since this is a static member we use a lock to prevent parallelism.
-			// This should be computed once only.
-			lock (_rowsPermutations)
-			{
-				if (!_rowsPermutations.TryGetValue(TargetSudokuGrid, out toReturn))
-				{
-					toReturn = GetRowsPermutationsUncached();
-					_rowsPermutations[TargetSudokuGrid] = toReturn;
-				}
-			}
-		}
-		return toReturn;
-	}
-
-	private IList<IList<IList<int>>> GetRowsPermutationsUncached()
-	{
-		var toReturn = new List<IList<IList<int>>>(9);
-		for (int i = 0; i < 9; i++)
-		{
-			var tempList = new List<IList<int>>();
-			foreach (var perm in AllPermutations)
-			{
-				// Permutation should be compatible with current row extended mask domains
-				if (Range9.All(j => ExtendedMask[(i, j)].Contains(perm[j])))
-				{
-					tempList.Add(perm);
-				}
-			}
-			toReturn.Add(tempList);
-		}
-
-		return toReturn;
-	}
-
-
-	/// <summary>
-	/// Produces 9 copies of the complete list of permutations
-	/// </summary>
-	public static IList<IList<IList<int>>> UnfilteredPermutations
-	{
-		get
-		{
-			if (!_unfilteredPermutations.Any())
-			{
-				lock (_unfilteredPermutations)
-				{
-					if (!_unfilteredPermutations.Any())
-					{
-						_unfilteredPermutations = Range9.Select(i => AllPermutations).ToList();
-					}
-				}
-			}
-			return _unfilteredPermutations;
-		}
-	}
-
-	/// <summary>
-	/// Builds the complete list permutations for {1,2,3,4,5,6,7,8,9}
-	/// </summary>
-	public static IList<IList<int>> AllPermutations
-	{
-		get
-		{
-			if (!_allPermutations.Any())
-			{
-				lock (_allPermutations)
-				{
-					if (!_allPermutations.Any())
-					{
-						_allPermutations = GetPermutations(Enumerable.Range(1, 9), 9);
-					}
-				}
-			}
-			return _allPermutations;
-		}
-	}
-
-	/// <summary>
-	/// The list of row permutations accounting for the mask
-	/// </summary>
-	public IList<IList<IList<int>>> TargetRowsPermutations
-	{
-		get
-		{
-			if (_targetRowsPermutations == null)
-			{
-				_targetRowsPermutations = GetRowsPermutations();
-			}
-			return _targetRowsPermutations;
-		}
-	}
-
-	/// <summary>
-	/// The list of compatible permutations for a given Sudoku is stored in a static member for fast retrieval
-	/// </summary>
-	private static readonly IDictionary<SudokuGrid, IList<IList<IList<int>>>> _rowsPermutations = new Dictionary<SudokuGrid, IList<IList<IList<int>>>>();
-
-	/// <summary>
-	/// The list of row indexes is used many times and thus stored for quicker access.
-	/// </summary>
-	private static readonly List<int> Range9 = Enumerable.Range(0, 9).ToList();
-
-	/// <summary>
-	/// The complete list of unfiltered permutations is stored for quicker access
-	/// </summary>
-	private static IList<IList<int>> _allPermutations = (IList<IList<int>>)new List<IList<int>>();
-	private static IList<IList<IList<int>>> _unfilteredPermutations = (IList<IList<IList<int>>>)new List<IList<IList<int>>>();
-
-	/// <summary>
-	/// Computes all possible permutation for a given set
-	/// </summary>
-	/// <typeparam name="T">the type of elements the set contains</typeparam>
-	/// <param name="list">the list of elements to use in permutations</param>
-	/// <param name="length">the size of the resulting list with permuted elements</param>
-	/// <returns>a list of all permutations for given size as lists of elements.</returns>
-	static IList<IList<T>> GetPermutations<T>(IEnumerable<T> list, int length)
-	{
-		if (length == 1) return list.Select(t => (IList<T>)(new T[] { t }.ToList())).ToList();
-
-		var enumeratedList = list.ToList();
-		return (IList<IList<T>>)GetPermutations(enumeratedList, length - 1)
-			.SelectMany(t => enumeratedList.Where(e => !t.Contains(e)),
-				(t1, t2) => (IList<T>)t1.Concat(new T[] { t2 }).ToList()).ToList();
-	}
-
-
 }
